Revert "Dw/pickup and throw"
