--- Assets/Animations.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 39104973506ee4545ab4a70c99e06195
-folderAsset: yes
-timeCreated: 1496509844
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
