--- Assets/Animations/HeadlessPete@Pickup.FBX.meta
@@ -1,229 +1,0 @@
-fileFormatVersion: 2
-guid: 72056e115d3334ccea4a41950ee7ea55
-timeCreated: 1496509683
-licenseType: Free
-ModelImporter:
-  serializedVersion: 19
-  fileIDToRecycleName:
-    100000: Bip01
-    100002: Bip01 Footsteps
-    100004: Bip01 Head
-    100006: Bip01 HeadNub
-    100008: Bip01 L Calf
-    100010: Bip01 L Clavicle
-    100012: Bip01 L Finger0
-    100014: Bip01 L Finger01
-    100016: Bip01 L Finger02
-    100018: Bip01 L Finger0Nub
-    100020: Bip01 L Finger1
-    100022: Bip01 L Finger11
-    100024: Bip01 L Finger12
-    100026: Bip01 L Finger1Nub
-    100028: Bip01 L Finger2
-    100030: Bip01 L Finger21
-    100032: Bip01 L Finger22
-    100034: Bip01 L Finger2Nub
-    100036: Bip01 L Finger3
-    100038: Bip01 L Finger31
-    100040: Bip01 L Finger32
-    100042: Bip01 L Finger3Nub
-    100044: Bip01 L Finger4
-    100046: Bip01 L Finger41
-    100048: Bip01 L Finger42
-    100050: Bip01 L Finger4Nub
-    100052: Bip01 L Foot
-    100054: Bip01 L Forearm
-    100056: Bip01 L Hand
-    100058: Bip01 L Thigh
-    100060: Bip01 L Toe0
-    100062: Bip01 L Toe01
-    100064: Bip01 L Toe02
-    100066: Bip01 L Toe0Nub
-    100068: Bip01 L UpperArm
-    100070: Bip01 Neck
-    100072: Bip01 Pelvis
-    100074: Bip01 R Calf
-    100076: Bip01 R Clavicle
-    100078: Bip01 R Finger0
-    100080: Bip01 R Finger01
-    100082: Bip01 R Finger02
-    100084: Bip01 R Finger0Nub
-    100086: Bip01 R Finger1
-    100088: Bip01 R Finger11
-    100090: Bip01 R Finger12
-    100092: Bip01 R Finger1Nub
-    100094: Bip01 R Finger2
-    100096: Bip01 R Finger21
-    100098: Bip01 R Finger22
-    100100: Bip01 R Finger2Nub
-    100102: Bip01 R Finger3
-    100104: Bip01 R Finger31
-    100106: Bip01 R Finger32
-    100108: Bip01 R Finger3Nub
-    100110: Bip01 R Finger4
-    100112: Bip01 R Finger41
-    100114: Bip01 R Finger42
-    100116: Bip01 R Finger4Nub
-    100118: Bip01 R Foot
-    100120: Bip01 R Forearm
-    100122: Bip01 R Hand
-    100124: Bip01 R Thigh
-    100126: Bip01 R Toe0
-    100128: Bip01 R Toe01
-    100130: Bip01 R Toe02
-    100132: Bip01 R Toe0Nub
-    100134: Bip01 R UpperArm
-    100136: Bip01 Spine
-    100138: Bip01 Spine1
-    100140: Bip01 Spine2
-    100142: Bip01 Spine3
-    100144: HeadlessPete
-    100146: //RootNode
-    400000: Bip01
-    400002: Bip01 Footsteps
-    400004: Bip01 Head
-    400006: Bip01 HeadNub
-    400008: Bip01 L Calf
-    400010: Bip01 L Clavicle
-    400012: Bip01 L Finger0
-    400014: Bip01 L Finger01
-    400016: Bip01 L Finger02
-    400018: Bip01 L Finger0Nub
-    400020: Bip01 L Finger1
-    400022: Bip01 L Finger11
-    400024: Bip01 L Finger12
-    400026: Bip01 L Finger1Nub
-    400028: Bip01 L Finger2
-    400030: Bip01 L Finger21
-    400032: Bip01 L Finger22
-    400034: Bip01 L Finger2Nub
-    400036: Bip01 L Finger3
-    400038: Bip01 L Finger31
-    400040: Bip01 L Finger32
-    400042: Bip01 L Finger3Nub
-    400044: Bip01 L Finger4
-    400046: Bip01 L Finger41
-    400048: Bip01 L Finger42
-    400050: Bip01 L Finger4Nub
-    400052: Bip01 L Foot
-    400054: Bip01 L Forearm
-    400056: Bip01 L Hand
-    400058: Bip01 L Thigh
-    400060: Bip01 L Toe0
-    400062: Bip01 L Toe01
-    400064: Bip01 L Toe02
-    400066: Bip01 L Toe0Nub
-    400068: Bip01 L UpperArm
-    400070: Bip01 Neck
-    400072: Bip01 Pelvis
-    400074: Bip01 R Calf
-    400076: Bip01 R Clavicle
-    400078: Bip01 R Finger0
-    400080: Bip01 R Finger01
-    400082: Bip01 R Finger02
-    400084: Bip01 R Finger0Nub
-    400086: Bip01 R Finger1
-    400088: Bip01 R Finger11
-    400090: Bip01 R Finger12
-    400092: Bip01 R Finger1Nub
-    400094: Bip01 R Finger2
-    400096: Bip01 R Finger21
-    400098: Bip01 R Finger22
-    400100: Bip01 R Finger2Nub
-    400102: Bip01 R Finger3
-    400104: Bip01 R Finger31
-    400106: Bip01 R Finger32
-    400108: Bip01 R Finger3Nub
-    400110: Bip01 R Finger4
-    400112: Bip01 R Finger41
-    400114: Bip01 R Finger42
-    400116: Bip01 R Finger4Nub
-    400118: Bip01 R Foot
-    400120: Bip01 R Forearm
-    400122: Bip01 R Hand
-    400124: Bip01 R Thigh
-    400126: Bip01 R Toe0
-    400128: Bip01 R Toe01
-    400130: Bip01 R Toe02
-    400132: Bip01 R Toe0Nub
-    400134: Bip01 R UpperArm
-    400136: Bip01 Spine
-    400138: Bip01 Spine1
-    400140: Bip01 Spine2
-    400142: Bip01 Spine3
-    400144: HeadlessPete
-    400146: //RootNode
-    4300000: HeadlessPete
-    7400000: Take 001
-    9500000: //RootNode
-    13700000: HeadlessPete
-  materials:
-    importMaterials: 1
-    materialName: 0
-    materialSearch: 1
-  animations:
-    legacyGenerateAnimations: 4
-    bakeSimulation: 0
-    resampleCurves: 1
-    optimizeGameObjects: 0
-    motionNodeName: 
-    rigImportErrors: 
-    rigImportWarnings: 
-    animationImportErrors: 
-    animationImportWarnings: 
-    animationRetargetingWarnings: 
-    animationDoRetargetingWarnings: 0
-    animationCompression: 1
-    animationRotationError: 0.5
-    animationPositionError: 0.5
-    animationScaleError: 0.5
-    animationWrapMode: 0
-    extraExposedTransformPaths: []
-    clipAnimations: []
-    isReadable: 1
-  meshes:
-    lODScreenPercentages: []
-    globalScale: 1
-    meshCompression: 0
-    addColliders: 0
-    importBlendShapes: 1
-    swapUVChannels: 0
-    generateSecondaryUV: 0
-    useFileUnits: 1
-    optimizeMeshForGPU: 1
-    keepQuads: 0
-    weldVertices: 1
-    secondaryUVAngleDistortion: 8
-    secondaryUVAreaDistortion: 15.000001
-    secondaryUVHardAngle: 88
-    secondaryUVPackMargin: 4
-    useFileScale: 1
-  tangentSpace:
-    normalSmoothAngle: 60
-    normalImportMode: 0
-    tangentImportMode: 3
-  importAnimation: 1
-  copyAvatar: 0
-  humanDescription:
-    serializedVersion: 2
-    human: []
-    skeleton: []
-    armTwist: 0.5
-    foreArmTwist: 0.5
-    upperLegTwist: 0.5
-    legTwist: 0.5
-    armStretch: 0.05
-    legStretch: 0.05
-    feetSpacing: 0
-    rootMotionBoneName: 
-    rootMotionBoneRotation: {x: 0, y: 0, z: 0, w: 1}
-    hasTranslationDoF: 0
-    hasExtraRoot: 0
-    skeletonHasParents: 1
-  lastHumanDescriptionAvatarSource: {instanceID: 0}
-  animationType: 2
-  humanoidOversampling: 1
-  additionalBone: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
