--- Assets/Materials/HeadlessPete.mat.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: e51895565cde04f988878581a788d160
-timeCreated: 1496510293
-licenseType: Free
-NativeFormatImporter:
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
