--- Assets/Materials/No Name.mat.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 016672773777d49988d2174c7b7818aa
-timeCreated: 1496509457
-licenseType: Free
-NativeFormatImporter:
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
