+++ Assets/Materials/background.mat
@@ -71,5 +71,5 @@ Material:
     - _UVSec: 0
     - _ZWrite: 1
     m_Colors:
-    - _Color: {r: 0.053471018, g: 0.16911763, b: 0.07819548, a: 1}
+    - _Color: {r: 0, g: 1, b: 0.21379328, a: 1}
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
