--- Assets/Meshes.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 8b7cfbb1a3bfc41778088c93a506e3a8
-folderAsset: yes
-timeCreated: 1496509822
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
