--- Assets/Meshes/HeadlessPete.FBX.meta
@@ -1,227 +1,0 @@
-fileFormatVersion: 2
-guid: aedf921883190411aa0ded79ae31a01f
-timeCreated: 1496509457
-licenseType: Free
-ModelImporter:
-  serializedVersion: 19
-  fileIDToRecycleName:
-    100000: Bip01
-    100002: Bip01 Head
-    100004: Bip01 HeadNub
-    100006: Bip01 L Calf
-    100008: Bip01 L Clavicle
-    100010: Bip01 L Finger0
-    100012: Bip01 L Finger01
-    100014: Bip01 L Finger02
-    100016: Bip01 L Finger0Nub
-    100018: Bip01 L Finger1
-    100020: Bip01 L Finger11
-    100022: Bip01 L Finger12
-    100024: Bip01 L Finger1Nub
-    100026: Bip01 L Finger2
-    100028: Bip01 L Finger21
-    100030: Bip01 L Finger22
-    100032: Bip01 L Finger2Nub
-    100034: Bip01 L Finger3
-    100036: Bip01 L Finger31
-    100038: Bip01 L Finger32
-    100040: Bip01 L Finger3Nub
-    100042: Bip01 L Finger4
-    100044: Bip01 L Finger41
-    100046: Bip01 L Finger42
-    100048: Bip01 L Finger4Nub
-    100050: Bip01 L Foot
-    100052: Bip01 L Forearm
-    100054: Bip01 L Hand
-    100056: Bip01 L Thigh
-    100058: Bip01 L Toe0
-    100060: Bip01 L Toe01
-    100062: Bip01 L Toe02
-    100064: Bip01 L Toe0Nub
-    100066: Bip01 L UpperArm
-    100068: Bip01 Neck
-    100070: Bip01 Pelvis
-    100072: Bip01 R Calf
-    100074: Bip01 R Clavicle
-    100076: Bip01 R Finger0
-    100078: Bip01 R Finger01
-    100080: Bip01 R Finger02
-    100082: Bip01 R Finger0Nub
-    100084: Bip01 R Finger1
-    100086: Bip01 R Finger11
-    100088: Bip01 R Finger12
-    100090: Bip01 R Finger1Nub
-    100092: Bip01 R Finger2
-    100094: Bip01 R Finger21
-    100096: Bip01 R Finger22
-    100098: Bip01 R Finger2Nub
-    100100: Bip01 R Finger3
-    100102: Bip01 R Finger31
-    100104: Bip01 R Finger32
-    100106: Bip01 R Finger3Nub
-    100108: Bip01 R Finger4
-    100110: Bip01 R Finger41
-    100112: Bip01 R Finger42
-    100114: Bip01 R Finger4Nub
-    100116: Bip01 R Foot
-    100118: Bip01 R Forearm
-    100120: Bip01 R Hand
-    100122: Bip01 R Thigh
-    100124: Bip01 R Toe0
-    100126: Bip01 R Toe01
-    100128: Bip01 R Toe02
-    100130: Bip01 R Toe0Nub
-    100132: Bip01 R UpperArm
-    100134: Bip01 Spine
-    100136: Bip01 Spine1
-    100138: Bip01 Spine2
-    100140: Bip01 Spine3
-    100142: HeadlessPete
-    100144: //RootNode
-    400000: Bip01
-    400002: Bip01 Head
-    400004: Bip01 HeadNub
-    400006: Bip01 L Calf
-    400008: Bip01 L Clavicle
-    400010: Bip01 L Finger0
-    400012: Bip01 L Finger01
-    400014: Bip01 L Finger02
-    400016: Bip01 L Finger0Nub
-    400018: Bip01 L Finger1
-    400020: Bip01 L Finger11
-    400022: Bip01 L Finger12
-    400024: Bip01 L Finger1Nub
-    400026: Bip01 L Finger2
-    400028: Bip01 L Finger21
-    400030: Bip01 L Finger22
-    400032: Bip01 L Finger2Nub
-    400034: Bip01 L Finger3
-    400036: Bip01 L Finger31
-    400038: Bip01 L Finger32
-    400040: Bip01 L Finger3Nub
-    400042: Bip01 L Finger4
-    400044: Bip01 L Finger41
-    400046: Bip01 L Finger42
-    400048: Bip01 L Finger4Nub
-    400050: Bip01 L Foot
-    400052: Bip01 L Forearm
-    400054: Bip01 L Hand
-    400056: Bip01 L Thigh
-    400058: Bip01 L Toe0
-    400060: Bip01 L Toe01
-    400062: Bip01 L Toe02
-    400064: Bip01 L Toe0Nub
-    400066: Bip01 L UpperArm
-    400068: Bip01 Neck
-    400070: Bip01 Pelvis
-    400072: Bip01 R Calf
-    400074: Bip01 R Clavicle
-    400076: Bip01 R Finger0
-    400078: Bip01 R Finger01
-    400080: Bip01 R Finger02
-    400082: Bip01 R Finger0Nub
-    400084: Bip01 R Finger1
-    400086: Bip01 R Finger11
-    400088: Bip01 R Finger12
-    400090: Bip01 R Finger1Nub
-    400092: Bip01 R Finger2
-    400094: Bip01 R Finger21
-    400096: Bip01 R Finger22
-    400098: Bip01 R Finger2Nub
-    400100: Bip01 R Finger3
-    400102: Bip01 R Finger31
-    400104: Bip01 R Finger32
-    400106: Bip01 R Finger3Nub
-    400108: Bip01 R Finger4
-    400110: Bip01 R Finger41
-    400112: Bip01 R Finger42
-    400114: Bip01 R Finger4Nub
-    400116: Bip01 R Foot
-    400118: Bip01 R Forearm
-    400120: Bip01 R Hand
-    400122: Bip01 R Thigh
-    400124: Bip01 R Toe0
-    400126: Bip01 R Toe01
-    400128: Bip01 R Toe02
-    400130: Bip01 R Toe0Nub
-    400132: Bip01 R UpperArm
-    400134: Bip01 Spine
-    400136: Bip01 Spine1
-    400138: Bip01 Spine2
-    400140: Bip01 Spine3
-    400142: HeadlessPete
-    400144: //RootNode
-    4300000: HeadlessPete
-    7400000: Take 001
-    9500000: //RootNode
-    13700000: HeadlessPete
-  materials:
-    importMaterials: 1
-    materialName: 0
-    materialSearch: 1
-  animations:
-    legacyGenerateAnimations: 4
-    bakeSimulation: 0
-    resampleCurves: 1
-    optimizeGameObjects: 1
-    motionNodeName: 
-    rigImportErrors: 
-    rigImportWarnings: 
-    animationImportErrors: 
-    animationImportWarnings: 
-    animationRetargetingWarnings: 
-    animationDoRetargetingWarnings: 0
-    animationCompression: 1
-    animationRotationError: 0.5
-    animationPositionError: 0.5
-    animationScaleError: 0.5
-    animationWrapMode: 0
-    extraExposedTransformPaths: []
-    clipAnimations: []
-    isReadable: 1
-  meshes:
-    lODScreenPercentages: []
-    globalScale: 1
-    meshCompression: 0
-    addColliders: 0
-    importBlendShapes: 1
-    swapUVChannels: 0
-    generateSecondaryUV: 0
-    useFileUnits: 1
-    optimizeMeshForGPU: 1
-    keepQuads: 0
-    weldVertices: 1
-    secondaryUVAngleDistortion: 8
-    secondaryUVAreaDistortion: 15.000001
-    secondaryUVHardAngle: 88
-    secondaryUVPackMargin: 4
-    useFileScale: 1
-  tangentSpace:
-    normalSmoothAngle: 60
-    normalImportMode: 0
-    tangentImportMode: 3
-  importAnimation: 1
-  copyAvatar: 0
-  humanDescription:
-    serializedVersion: 2
-    human: []
-    skeleton: []
-    armTwist: 0.5
-    foreArmTwist: 0.5
-    upperLegTwist: 0.5
-    legTwist: 0.5
-    armStretch: 0.05
-    legStretch: 0.05
-    feetSpacing: 0
-    rootMotionBoneName: Bip01
-    rootMotionBoneRotation: {x: 0, y: 0, z: 0, w: 1}
-    hasTranslationDoF: 0
-    hasExtraRoot: 0
-    skeletonHasParents: 1
-  lastHumanDescriptionAvatarSource: {instanceID: 0}
-  animationType: 2
-  humanoidOversampling: 1
-  additionalBone: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
