--- Assets/Prefabs/Body.prefab.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 0ca569d03ea92db44a89cdd48433ae43
-timeCreated: 1496515176
-licenseType: Free
-NativeFormatImporter:
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
