+++ Assets/Prefabs/Head.prefab
@@ -22,11 +22,11 @@ GameObject:
   - component: {fileID: 33000012384968284}
   - component: {fileID: 135000010006721558}
   - component: {fileID: 23000012558626572}
+  - component: {fileID: 114000013718454806}
   - component: {fileID: 54000010978611206}
-  - component: {fileID: 114367429796193580}
   m_Layer: 0
   m_Name: Head
-  m_TagString: HeadIsDetached
+  m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
@@ -71,7 +71,7 @@ Transform:
   m_GameObject: {fileID: 1000010073205616}
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
   m_LocalPosition: {x: 0, y: 0.5, z: 0}
-  m_LocalScale: {x: 0.5, y: 0.5, z: 0.5}
+  m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
   - {fileID: 4000011601553612}
   m_Father: {fileID: 0}
@@ -181,7 +181,7 @@ Behaviour:
   m_PrefabInternal: {fileID: 100100000}
   m_GameObject: {fileID: 1000011871868080}
   m_Enabled: 1
---- !u!114 &114367429796193580
+--- !u!114 &114000013718454806
 MonoBehaviour:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
@@ -189,7 +189,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1000010073205616}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3a818b8e86597ce429469e043889381f, type: 3}
+  m_Script: {fileID: 11500000, guid: 66e3b352e3366f54b89371783b04d472, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   speed: 5
