+++ Assets/Scripts/BodyController.cs
@@ -4,14 +4,13 @@ using System.Collections;
 public class BodyController : MonoBehaviour
 {
 
-	public float movementSpeed;
-	public float rotationSpeed;
+    public float speed;
 
     // private Rigidbody rb;
 
     void Start()
     {
-//        Rigidbody rb = GetComponent<Rigidbody>();
+        Rigidbody rb = GetComponent<Rigidbody>();
     }
 
     void FixedUpdate()
@@ -22,8 +21,6 @@ public class BodyController : MonoBehaviour
         // Vector3 movement = new Vector3(moveHorizontal, 0.0f, moveVertical);
         //
         // rb.AddForce(movement * speed);
-		transform.Translate(0f, -(movementSpeed * Input.GetAxis("Vertical2") * Time.deltaTime), 0f);
-		transform.Rotate(0f, 0f, rotationSpeed * Input.GetAxis("Horizontal2") * Time.deltaTime);
-		
+		transform.Translate(speed*Input.GetAxis("Horizontal2")*Time.deltaTime, 0f, speed*Input.GetAxis("Vertical2")*Time.deltaTime);
     }
 }
