--- Assets/Scripts/HeadController.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 3a818b8e86597ce429469e043889381f
-timeCreated: 1496512773
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
