--- Assets/Scripts/PickupHead.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 25fb31c873aa17b4e81a9b3e717bedb2
-timeCreated: 1496512612
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
