+++ Assets/Scripts/PlayerController.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 66e3b352e3366f54b89371783b04d472
+timeCreated: 1496457059
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
