--- Assets/Scripts/ThrowHead.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 5136ce27eb448b543998d3bf6185e6fa
-timeCreated: 1496514639
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
