--- Assets/Textures.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: cf0fb4471d74942128123e7c961f3000
-folderAsset: yes
-timeCreated: 1496509836
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
