+++ ProjectSettings/TagManager.asset
@@ -3,10 +3,7 @@
 --- !u!78 &1
 TagManager:
   serializedVersion: 2
-  tags:
-  - Body
-  - HeadIsAttached
-  - HeadIsDetached
+  tags: []
   layers:
   - Default
   - TransparentFX
